+++ plugins/calendar/src/test/java/de/unistuttgart/iaas/amyassist/amy/plugin/calendar/CalendarLogicTest.java
@@ -53,7 +53,7 @@ public class CalendarLogicTest {
 	 */
 	@BeforeEach
 	public void setup() {
-		this.callog = this.framework.setServiceUnderTest(CalendarLogic.class);
+		// this.callog = this.framework.setServiceUnderTest(CalendarLogic.class);
 	}
 
 	/**
